LocalFirstAuthProvider: Adding implementation for addDevice
I've used the getShare function from previous commits to get the share and get the share.team.members to push a device. I faced an issue of devices aray coming as undefined. I added a little hack to keep making progress but it sounds like I've cheated here so we will probably need to change it later.