LocalFirstAuthProvider: Adding implementation for getState
I've used the loadTeam func but it was asking for the team graph that I got from share.team.graph and using the getShare func as well and the team.teamKeys() when we are creating the team. We will need to think in a better solution for when we will have more than one share for an auth provider but at least now the test is passing.